Temp resolved image off center bug in mid section
Vertical flexboxes in mid section depend are hard-coded values to match
the width of the images in the section. Once the skeleton of the site
is laid out, this needs to be revisted. Flex grow/shrink using the
keyword 'auto' comes to mind right now. But this will need to be looked
at in the future.